Remove all reference to the client part of the project
To get this to build the project needs more disk space and memory than we get with a "free tier" AWS VM. The big problem is the Angular component, which also adds a _lot_ of complication (Caddy, etc.).

So I'm just going to strip all that out and we'll instead build just the server and MongoDB containers.